+++ CHANGELOG.md
@@ -9,7 +9,10 @@ Changelog follow https://keepachangelog.com/ format.
 ## [Unreleased]
 
 *   `epath`: Add `missing_ok=False` kwargs to `path.rmtree`.
-*   `enp`: Fix `ArraySpec.from_array` when TF is in graph mode.
+*   `enp`:
+    *   `ArraySpec.from_array`:
+        *   Fix when TF is in graph mode.
+        *   Add `orbax` metadata support
 *   `epy`:
     *   Add `epy.pretty_repr` for pretty print an object (including dataclass).
     *   Add `epy.diff_str` for pretty diff print of 2 objects.