+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Changelog follow https://keepachangelog.com/ format.
 
 ## [Unreleased]
 
+## [1.5.0] - 2023-09-19
+
 *   `ecolab`:
     *   Auto display statements ending with `;` (assignments, return
         statements, expressions,...).
@@ -312,8 +314,9 @@ Changelog follow https://keepachangelog.com/ format.
 
 *   Automated github release
 
-[Unreleased]: https://github.com/google/etils/compare/v1.4.1...HEAD
-[1.4.0]: https://github.com/google/etils/compare/v1.4.0...v1.4.1
+[Unreleased]: https://github.com/google/etils/compare/v1.5.0...HEAD
+[1.5.0]: https://github.com/google/etils/compare/v1.4.1...v1.5.0
+[1.4.1]: https://github.com/google/etils/compare/v1.4.0...v1.4.1
 [1.4.0]: https://github.com/google/etils/compare/v1.3.0...v1.4.0
 [1.3.0]: https://github.com/google/etils/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/google/etils/compare/v1.1.1...v1.2.0